RIDCP_dehazing_MindSpore-master/requirements.txt
addict
future
lmdb
numpy
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
# torch>=1.7
torch==1.13.1
torchvision
tqdm
yapf
pyiqa>=0.1.4
einops
mindspore-dev
mindcv